backend/FasterRCNN/requirements.txt
albucore               0.0.23
albumentations         2.0.5
annotated-types        0.7.0
certifi                2025.1.31
charset-normalizer     3.4.1
colorama               0.4.6
contourpy              1.3.1
cycler                 0.12.1
filelock               3.13.1
fonttools              4.56.0
fsspec                 2024.6.1
idna                   3.10
Jinja2                 3.1.4
kiwisolver             1.4.8
MarkupSafe             2.1.5
matplotlib             3.10.1
mpmath                 1.3.0
networkx               3.3
numpy                  2.1.2
opencv-python          4.11.0.86
opencv-python-headless 4.11.0.86
packaging              24.2
pillow                 11.0.0
pip                    24.2
pycocotools            2.0.8
pydantic               2.11.1
pydantic_core          2.33.0
pyparsing              3.2.1
PyQt5                  5.15.11
PyQt5-Qt5              5.15.2
PyQt5_sip              12.17.0
python-dateutil        2.9.0.post0
PyYAML                 6.0.2
requests               2.32.3
scipy                  1.15.2
setuptools             70.2.0
simsimd                6.2.1
six                    1.17.0
stringzilla            3.12.3
sympy                  1.13.1
torch                  2.1.0+cu118
torchaudio             2.6.0+cu118
torchvision            0.21.0+cu118
tqdm                   4.67.1
typing_extensions      4.12.2
typing-inspection      0.4.0
urllib3                2.3.0
--extra-index-url https://download.pytorch.org/whl/cu118